Dodano optymalizację dystansu z domu kransala
